refactor - http 관련 클래스들을 http 패키지 하위로 옮김
