+++ src/main/java/webserver/http/ProtocolVersion.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.http;
 
 import java.util.Objects;
 import java.util.regex.Matcher;
