+++ src/main/java/webserver/http/QueryString.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.http;
 
 import java.util.Collections;
 import java.util.Map;
