+++ src/main/java/webserver/http/RequestLine.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.http;
 
 public class RequestLine {
