+++ src/test/java/webserver/http/ProtocolVersionTest.java
@@ -1,8 +1,9 @@
-package webserver;
+package webserver.http;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import webserver.http.ProtocolVersion;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
